Fixes bug with Lua
The 'find' command takes a pattern with no way to explicitly inform it 
that you might want to find something literal.

This is a bug with the specification of Lua, which has required me to borrow
a 'litpat' function from the aethyr. We now call this to avoid blowing up
when looking for characters like "(".