Completely bogus 'fix'
This only 'works' because it breaks the paradigm, I'm
going to need to apply some serious brain muscle to figure out 
how to actually make this work. 

It may be as simple as making ts, not tabulate, the function which
repeatedly calls lineGen until there's nothing left. 

ts_coro and _tabulate (you can tell these are legacy because the names
suck) are co-recursive, so this might work. Hmm.

I think it's reasonable to finish fixing up the logic so that it indents
properly and detects long lines, then go back and make it work correctly.