Acceptable
This works.... okay.

It's stuck with a trailing comma and closing braces on their own line,
for now. 

I don't see any good way to prevent that before I tighten up the use of widths.

Right now, I'm eager to wire this into rainbuf so that we, at long last,
get JIT printing of values.