Make a clear note of the issues with __G, os.exit(), etc.
Move cleanup code back to after uv.run() (out of the ^Q handler).
Add comments explaining the need for immediate evaluation of the module and the os.exit() call.
Save and restore the cursor using terminal escapes specifically for this purpose, with a note that this may not work everywhere and may need to be replaced.