Use custom environment for evaluation
With this refinement, we now only eval the Txtbuf once per eval call.

In the previous system, we run the chunk up to three times, with
unintuitive results in the case of any side effects.

This also adds any new globals to the repr anti_G table, allowing the
results to keep up with new information.