Reorder parameters of tabulate and remove extraneous params from lineGen
lineGen doesn't use or need to know anything about `depth` or `cycle`, so it doesn't need them as params and *could* just pass nils to the invocation of tabulate.
But really, since those params are optional and `phrase` is not, `phrase` should be second in the arg list and then lineGen can just call with two params, leaving the others nil by default.