Make fewer assumptions about the content of separator tokens
Allow them to end with a non-space character, or contain multiple non-space characters.

Add a removeTrailingSpaces() method to Token and use that instead of blindly removing the last character
Check the length of the line even when hitting a separator, but keep in mind that we don't need to reserve space after it since we'll be wrapping anyway.