Localize vars in token.orb and repr.orb
This makes `helm/repr` callable from outside of helm.

I need to decide what I want to do about values that should be in
the global namespace.

Originally, and I'll probably do this, 'core' was going to be a
module that was loaded inside bridge, and which puts things into
the global namespace and adds things to the core tables.

This branch isn't intended for merging, it's just something I'm
using while I work on option flags for `br orb`.