Use weak keys on anti_G table
Previously, when making an anti_G table, the reference stored in
the table would prevent anything in the namespace from ever being
garbage collected.

This being suboptimal, we change anti_G to be a weak-key table,
so that any object which is no longer referenced elsewhere will be
collected.  This will come in handy when it comes time to start
doing hot reloads, and is intrinsically the correct behavior.