Add reloading (but not rerunning) with ^R
It's possible that this is leaking an entire global environment
each time I run it, but, I don't think so; _G and the __G
metatable should be the only references to the global table.

bridge is already running under __G and uses the global _G to
reset things (although we immediately exit after that, it seems
polite).

I dunno. Environments are weird and I'm a lil bit paranoid...