Persist all iterations of session
Eventually, when we make the format more sophisticated, we can do
a comparison between sessions.

For now I think this is a good compromise. It doesn't leave
redundant copies of the results and Txtbufs sitting on the history,
but it preserves differences in the database for later analysis and
retrieval.

The alternatives are to only keep the first, which seems bad, and
to delete results, which also seems bad.

This also rewrites `persist` so it can run as a blocking function.

This may not be necessary; as long as we write the line_ids in the
proper order, we should be okay, and I could use the savepoint
mechanism to make named transactions that should in principle all
roll back if the program crashes.

But I don't know what happens if savepoints overlap in the same
conn, and don't have the time to find out. This is okay for now.