Actually fetch latest line_id of project
The previous version had a race condition where another instance
of helm using the same database could commit a line between the
insert and the latestRowId, which would give the wrong results.

However unlikely that might have been, this at least reduces that
possibity to two helm repls, running in the same project.

It may even eliminate it, because it's inside a transaction.