Explicitly stop timer
Best to clean up resources, I don't know what happens to the timer
in the old version, while I *do* know what happens in the new
version.

Also asserted os.time and math.random into historian; it was 'good'
that I hadn't, because it surfaced a bug in deepclone(), but it's
still the right thing to do.