Move seq dispatch logic to the raga
Extract an abstract RagaBase that pretty much just handles that for now, and an EditBase for ragas that are primarily about editing the Txtbuf. This is all of them for now, but seems like a useful distinction to make.
Add an action_complete property to ModeS to allow a raga to say "okay, retry that please".
Defer switching ragas until after seq dispatch is complete by adding a shift_to property to ModeS.