Refactor left and right delta into 'delta'
This is fully general.

Caveat: The way this is written, the 'default' direction is
backward, if delta is called with two parameters.

But I thought it was more confusing if backward was the default in
the code, so this was the compromise.

There's no actual reason to call delta with two parameters, it just
happens to work if one does.