Keep package.loaded, remove packages on restart
`package.loaded` is accessed from the registry when `require`
checks it.  The reference in _G is a convenience; replacing it
does not replace the "one true" `package.loaded`.

That's aesthetically unsatisfying, but we have a workaround: we
now cache a Set of the keys in `package.loaded` when we first run
`modeS:eval()`, and take the difference between the current state
of packages and that original state, removing any packages which
have been added at the repl.

The end result is that if a `require`ed module is changed, restart
will reflect that change.

This is an important step toward a "true" REPL, the kind that
doesn't get restarted very often.