Add profiler at top of helm
This shows that, in fact, we spend the bulk of time on stepping
through the load history buffer (at least for projects like helm
where there is significant history to load).

I *think* this happens off the main loop, in pieces. It's certainly
intended to be that way.

But I haven't found a place to drop `profile.stop()` where it
doesn't spend a lot of time in `_step`.  So it's possible we're
doing something wrong here, even if the most likely wrong thing is
where I've put the profile stop code thus far.