Rewrite migration 3 as table of SQL statements
This is less eloquent than I would like.

It allows the use of native sql blocks, which is good; but it uses
numeric indices, which is brittle.

I especially dislike that I have Lua blocks interspersed with
SQL blocks, but this, at least, is a straightforward substitute
for transclusion, which we can't do yet

One improvement would be to highlight names inside codeblock
headers, in the Orb.sublime-syntax file.

In any case, it's better than the 'inner language' effect we get
with strings.  It will do for now.