Move session info to session table on modeS
Realizing I have to pass session info into historian to take care
of persistence, and that the modeselektor will need that info
more generally for interactive stuff (so we don't want to put it
on the historian instead).

So better to put it inside its own table, than pass the entire
modeS into historian.  I don't think historian has any hard
dependencies on a modeselektor existing, and I certainly would not
want to add one.