Write out migration for result table
Untested.

The plan is to run the simple migrations first, check the results
with sqlite3, and see if any changes in the application code are
necessary to prevent show-stopping regressions.

Then we'll try out the result migration, and see if the shape of
the database is correct. This will definitely break app code.