Restore pre-migration-6 helm-db order
Having added a copy of the schema, it better serves the purpose of
being a canonical reference for the database.

So before I merge this, I'll take a pass to put the tables into
the migrations, so they can serve as a story of how the database
came to be the way that it is.

We can optionally just skip running migrations for new installs,
and use the schema directly to recreate the database, although I'm
wary of doing this, lest the migrations themselves suffer bitrot.