Update to new `valiant` API
Treat the valiant:valiant module as a class—capitalize the constructor name
Expect only the one return value from the Valiant constructor and call :restart()/:reset() against it
Accept a `Valiant` instance in `Session:evaluate()` rather than a boolean telling it whether to "isolate" itself or not. Note this is unused right now, so no changes to call sites. Note also that for the moment we can no longer isolate anti_G, but I expect to fix this by having Valiant manage that as well.