Respect the width of the gutter when rendering a Resbuf
Somewhere along the way I refactored so what you pass to :initComposition() may not be the same as what is stored in .cols, in order to account for the gutter. Overrides of :initComposition() need to not access the argument directly once they've called super as a result. This is...less than great, really we should probably save what we were given and have a method that adjusts it on the fly, but I don't want to spend more time on this right now.