Clear caches when the width of a Rainbuf changes
I feel like we've gone back and forth on this a bit, it was this way at one point, then seems to have stopped being, not sure why. Ultimately it's pretty clear to me that Rainbuf really does need to know its width in order to work right, and if you want multiple viewports on the same thing you're going to have to make multiple Rainbufs. (Actually it also needs to know its _height_, because scrolling pretty clearly should be handled by Rainbuf as well, but that's another story.)