Use a consistent emoji for 'reject'
We use the red cross for a failed premise in session reporting, and
a red no for a rejected premise.

This is already moderately confusing, but I don't see an obviously
better iconography.

In any case, the emoji should be consistent between the two
projects. A check for yes and and x for no are well-established
conventions, and the "no" sign suggests deliberate failure, so
this is the correct pairing.