Disallow starting a new session with the same title as an existing one using -n or -M
Make the Session() constructor accept the title-or-index as a parameter and try to load, falling back to create if that fails. Then the calling code in Historian() can check if what we got matches our expectations and throw an appropriate error if not.

Depends on argparse in `pylon` disallowing the case of "hey make me a session named '3'".