First stab at an EditAgent
Move 95% of Txtbuf code to EditAgent
Txtbuf follows .value convention like other Rainbufs
Can remove some overrides to avoid stomping the results--it's now okay to clear the ResultsAgent more aggressively when the results Zone is being supplied by a different agent
Remove :setTxtbuf(), the actual Txtbuf is never replaced
Suggestions and lexing still live on Txtbuf for now, but it's becoming increasingly clear that they should live on EditAgent. Suggestions are sorta broken right now as a result, fixing that is the next step.