Use a queue for results idler
We were getting some resource exhaustion and race condition issues
with the old way of doing things.

This does require getting rid of the savepoints, and if we wanted
that back we'll have to think up a different way of handling the
transaction.

I don't think the risk of losing the occasional result in a hard
crash is big enough that we should care about that right now.