Move keymaps to a keymap/ directory, making them static/constructed once at load time
No more keymap fragments on Agents or Ragas.

Use `target =` instead of `source =` in keymap declaration—the bindings are no longer being retrieved from the source, so it's *not* a source of anything.

Move logic to begin session review on quit into modeS:quit(). Not a good final resting place—or maybe it just needs a better name—but gets it off RagaBase. Overridden quit logic can now be just additional keymap entries that call a different method.

Move :openHelpOnFirstKey to modeS—should probably have a HelpAgent or something eventually?