Basic interactive-restart implementation
Review raga is now a base implementation for RunReview and SessionReview
Extract common review keybindings to keymap/parts.review_common, this approach isn't sustainable though.
Add EditLine raga—right now only for editing a to-be-inserted line, but I'd *like* to use it for in-place edits too, not sure that'll be quite possible though
Rename Sessionbuf -> Reviewbuf, turns out not to need specialization though
Add RunReviewAgent
Rename "skip" to "trash" in session review for consistency with run review and the icon. (This is never shown to the user as text nor persisted to the database.)