Insert new premises/lines before the current selection, rather than after
Add a "was_inserting" flag so we don't get stuck going "keep" -> [Tab] adds a new line that is "insert" -> [Tab] removes that line and the old one is still on "keep" -> [Tab] adds a new line. Instead, IFF we just *left* insert, it will be skipped in the order so the next [Tab] changes the selected, preexisting premise to "trash".
Add a :selectionChanged() hook up on ReviewAgent which RunReviewAgent overrides to help manage the flag.