Cache results by line_id rather than cursor index on Historian
Extract a methods Historian:resultsFor(line_id) to allow other agents to retrieve results for arbitrary lines (e.g. for Sessions)—moving towards Historian having sole responsibility for results.
Don't append to hist.line_ids in :persist, but *do* cache the result in the result_buffer. :append is the proper place to manipulate .line_ids as it must be kept parallel with the indexed fields of Historian itself.