Keep lines and their results together as Rounds
Still on Historian, but no more .line_ids or .result_buffer
Introduce the concept of the "desk"—the round which is being actively edited and has not yet been evaluated. It occupies a pseudo-index of historian.n + 1, but is not actually added to the indexed slots until evaluated.
Simplify stashing of edits-in-progress—just always clobber the desk, don't actually append to the history until a line is evaluated. This fixes getting tons of dupes in the history when just scrolling around, while allowing successive re-evaluations of the same line to still appear in the history.
Add Round order with line and response. The response is wrapped in a single-element array table, with a :results() method that returns the actual results. Database results are additionally kept as .db_result. Further refinement needed here for sure.
Remove {old|new}_{result|line_id}, etc. from session and run-review premises. Replace with, for now, a single `.round` property, which can easily be replaced by a .{old|new}_round pair as needed.
Remove Session:resultsAvailable(), the Round just gets modified in place once result persistence is done. This is maybe not ideal (shared state), but if so, we'll need an Action, so :resultsAvailable would go away anyway.